move update_argument_collection into argument.py
